+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/tweag/gomod2nix/cmd"
+import "github.com/tweag/gomod2nix/internal/cmd"
 
 func main() {
 	cmd.Execute()
